Add custom limits/requests for the audit controller
Some of the most resource intensive functions of gatekeeper have been split off
into the audit controller, so it's deserving of some of the custom defaults we
established when the webhook and the audit were ran in the same pod.

I'm only limiting the audit to 2000m, rather than 7000m because it doesn't have
the same potential to slow down requests on the api server that the webhook does
when its throttled.