Set audit-from-cache to true (#23)
By default the gatekeeper audit will use the k8s discovery client to discover
objects based on the selectors present on the constraint. This is a lot more CPU
and memory intensive than the previous default which used the objects cached for
use in OPA policies, which are synced according to a predefined list of resource
types.

As we're only using a small, known subset of resource types for our constraints,
I think it's preferable to use the OPA cache.