Adjust audit CPU/memory limits
The controller serves the webhook which has the potential to
dramatically impact apiserver performance if it was OOM killed or CPU
throttled, hence the high limits.

The audit manager on the other hand could be throttled/OOM killed
without affecting cluster functionality and therefore doesn't
necessarily justify the sky high limits.